backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI/ML
google-genai==1.0.0
numpy==1.26.3
scikit-learn==1.4.0

# Storage & Files
boto3==1.34.25
python-magic==0.4.27
Pillow==10.2.0

# Utilities
httpx==0.26.0
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Websockets
websockets>=13.0
